Async Flattening
Now the function return scope is relative to when the actual task completes, while still allowing it to be async under the hood. It's so cool how much you can get done when using async code, so awesome! Not that inherently changes how the code runs, but if it is a heavy task, and you lift it off of being blocking, then it definitely helps to run them at the same time.