Module Features
Now the module allows external projects to access this setup.

I wanted to make it so you could pass `--test` when running this module, but I didn't want to rely on weird args parsing just for this. Usually I put my tests in a separate file, and run the main module by using it's exports like as this commit adds. Once we have more test files available to use, I can help setup a standalone unit test file for that too.